Fix partition resolve bug for materialized view
